Added additional non-backticks style versions
Pretty minor tweak, this is how react handle's these 2 class / for, I figured I would add typeName as well for consistency.

className and htmlFor is how the DOM API and ReactJs handle this. I thought matching the DOM API was a pretty decent way to go. Feel free to reject if it's you don't like the extra verbosity.