Refine text tags (#232)
* Refine tags to a more specific type

There is a regression introduced during the Scala 3 upgrade where the
`scalatags.Text.all.doctype` case class no longer worked with the
pre-existing tags since its second argument accepts a
`scalatags.text.Frag` but the tags themselves refined to
`<tag-name>.Self` (the <tag-name> is a placeholder for the tag's name).
As it turns out, this self type was not being refined by the tags
themselves because for some reason, the Scala 3 compiler resolves this
type to be `scalatags.generic.TypedTag` i.e. it resolves to the more
generic type as opposed to the more specific type of
`scalatags.Text.TypedTag`. Fortunately in Scala 2, this works out fine
as the type does get refined and therefore the tags can be used inside
a `scalatags.text.Frag`.

This commit essentially adds type annotations to all the tags under the
text package so that the more specific `ConcreteHtmlTag[String]` is used
instead of the more generic `TypedTag`.

* Add doctype compilation test and fix extraneous warnings.