update various deps and drop scalajs 0.6 (#239)
The original reason for this was just to update the com-lihaoyi name so
that when updates come from Scala Steward they correctly link back to
this repo. However I also then bumped a bunch of outdated stuff. As a
follow-up to the conversation on Discord I also removed publishing for
0.6.33. If we don't, then we can't update the latest versions of each
Scala. As far as I can tell, no one is still publishing for 0.6.x and
when I asked on the ScalaJS channel it was pretty clear that it's not
necessary. So I can't find a good reason to keep doing it, but let me
know. Arguably you can do the same for Scala 2.11, but I won't press
that button :).