docs: make reference to stylesheet `val`s clear
In Scala 2 you could use a `def` to define your `cls`s, but in Scala 3
the implementation is only looking for `valDef`s in the file, meaning
that `def`s will no longer work. This makes sure the docs don't mention
using `def`s and instead focus the verbiage on `val`s.