refactor: allow for `customSheetName` to be an empty string
While you could do this before, if you had a class `a` and you wanted to
keep it simple and not have anything for your `customSheetName` you'd
end up with `-a` which was a bit annoying. This small change allows
users to still use this the way the were before, but now if they are
using an empty string as their `customSheetName`, they'll just get `a`
instead of `-a`.